// extension/pplcz/src/Admin/MuiAdmin/src/schema.ts
export interface components {
  schemas: {
// ...
    ShipmentWithAdditionalModel: {
      shipment: components["schemas"]["ShipmentModel"];
      errors?: components["schemas"]["WpErrorModel"][] | null;
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}